feat: add jvm (desktop) and wasmjs (web) target, also make demo app multiplatform
