+++ app/build.gradle
@@ -1,6 +1,61 @@
+import org.jetbrains.kotlin.gradle.tasks.KotlinJvmCompile
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
+
 plugins {
     id 'com.android.application'
-    id 'kotlin-android'
+    id "kotlin-multiplatform"
+    id "org.jetbrains.compose"
+}
+
+kotlin {
+    applyDefaultHierarchyTemplate()
+
+    androidTarget()
+    jvm("desktop")
+    wasmJs {
+        moduleName = "capturable-demo"
+        browser {
+            commonWebpackConfig {
+                outputFileName = "composeApp.js"
+            }
+        }
+        binaries.executable()
+    }
+
+    sourceSets {
+        commonMain {
+            dependencies {
+                implementation(compose.runtime)
+                implementation(compose.foundation)
+                implementation(compose.material)
+                implementation(compose.ui)
+                implementation(compose.components.resources)
+
+                // Capturable library
+//    implementation "dev.shreyaspatil:capturable:2.1.0"
+                implementation(project(":capturable"))
+            }
+        }
+        androidMain {
+            dependencies {
+                // Android
+                implementation 'androidx.core:core-ktx:1.12.0'
+                implementation 'androidx.appcompat:appcompat:1.6.1'
+                implementation 'com.google.android.material:material:1.11.0'
+                implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.7.0'
+                implementation 'androidx.activity:activity-compose:1.8.2'
+            }
+        }
+        desktopMain {
+            dependencies {
+                implementation(compose.desktop.currentOs)
+            }
+        }
+    }
+
+    compilerOptions {
+        optIn.add("org.jetbrains.compose.resources.ExperimentalResourceApi")
+    }
 }
 
 android {
@@ -29,9 +84,6 @@ android {
         sourceCompatibility JavaVersion.VERSION_11
         targetCompatibility JavaVersion.VERSION_11
     }
-    kotlinOptions {
-        jvmTarget = '11'
-    }
     buildFeatures {
         compose true
     }
@@ -44,26 +96,24 @@ android {
         }
     }
     namespace 'dev.shreyaspatil.capturableExample'
+    dependencies {
+        // Jetpack Compose
+        implementation platform("androidx.compose:compose-bom:2024.04.00")
+        implementation "androidx.compose.ui:ui-tooling-preview"
+        debugImplementation "androidx.compose.ui:ui-tooling"
+    }
 }
 
-dependencies {
+compose.desktop {
+    application {
+        mainClass = "MainKt"
+    }
+}
 
-    // Capturable library
-//    implementation "dev.shreyaspatil:capturable:2.1.0"
-    implementation(project(":capturable"))
+compose.web {}
 
-    // Android
-    implementation 'androidx.core:core-ktx:1.12.0'
-    implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'com.google.android.material:material:1.11.0'
-    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.7.0'
-    implementation 'androidx.activity:activity-compose:1.8.2'
-
-    // Jetpack Compose
-    implementation platform("androidx.compose:compose-bom:2024.04.00")
-    implementation "androidx.compose.foundation:foundation"
-    implementation "androidx.compose.ui:ui"
-    implementation "androidx.compose.material:material"
-    implementation "androidx.compose.ui:ui-tooling-preview"
-    debugImplementation "androidx.compose.ui:ui-tooling"
+tasks.withType(KotlinJvmCompile.class).configureEach {
+    compilerOptions {
+        jvmTarget.set(JvmTarget.JVM_11)
+    }
 }
