+++ app/src/androidMain/res/drawable/ic_baseline_qr_code_24.xml
@@ -5,36 +5,36 @@
     android:viewportHeight="24"
     android:tint="?attr/colorControlNormal">
   <path
-      android:fillColor="@android:color/white"
+      android:fillColor="#FFFFFF"
       android:pathData="M3,11h8V3H3V11zM5,5h4v4H5V5z"/>
   <path
-      android:fillColor="@android:color/white"
+      android:fillColor="#FFFFFF"
       android:pathData="M3,21h8v-8H3V21zM5,15h4v4H5V15z"/>
   <path
-      android:fillColor="@android:color/white"
+      android:fillColor="#FFFFFF"
       android:pathData="M13,3v8h8V3H13zM19,9h-4V5h4V9z"/>
   <path
-      android:fillColor="@android:color/white"
+      android:fillColor="#FFFFFF"
       android:pathData="M19,19h2v2h-2z"/>
   <path
-      android:fillColor="@android:color/white"
+      android:fillColor="#FFFFFF"
       android:pathData="M13,13h2v2h-2z"/>
   <path
-      android:fillColor="@android:color/white"
+      android:fillColor="#FFFFFF"
       android:pathData="M15,15h2v2h-2z"/>
   <path
-      android:fillColor="@android:color/white"
+      android:fillColor="#FFFFFF"
       android:pathData="M13,17h2v2h-2z"/>
   <path
-      android:fillColor="@android:color/white"
+      android:fillColor="#FFFFFF"
       android:pathData="M15,19h2v2h-2z"/>
   <path
-      android:fillColor="@android:color/white"
+      android:fillColor="#FFFFFF"
       android:pathData="M17,17h2v2h-2z"/>
   <path
-      android:fillColor="@android:color/white"
+      android:fillColor="#FFFFFF"
       android:pathData="M17,13h2v2h-2z"/>
   <path
-      android:fillColor="@android:color/white"
+      android:fillColor="#FFFFFF"
       android:pathData="M19,15h2v2h-2z"/>
 </vector>
