+++ app/src/androidMain/res/mipmap-mdpi/ic_launcher.webp
