+++ app/src/desktopMain/kotlin/Main.kt
@@ -1,0 +1,11 @@
+import androidx.compose.ui.window.Window
+import androidx.compose.ui.window.application
+import dev.shreyaspatil.capturableExample.ui.App
+
+fun main() = application {
+    Window(
+        onCloseRequest = ::exitApplication
+    ) {
+        App()
+    }
+}
