+++ app/src/wasmJsMain/kotlin/main.kt
@@ -1,0 +1,12 @@
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.window.CanvasBasedWindow
+import dev.shreyaspatil.capturableExample.ui.App
+
+@OptIn(ExperimentalComposeUiApi::class)
+fun main() {
+    CanvasBasedWindow(
+        canvasElementId = "ComposeTarget"
+    ) {
+        App()
+    }
+}
