+++ gradle.properties
@@ -33,3 +33,4 @@ kotlin.mpp.androidGradlePluginCompatibility.nowarn=true
 
 #Compose
 org.jetbrains.compose.experimental.uikit.enabled=true
+org.jetbrains.compose.experimental.wasm.enabled=true
